feat(fsd): A2 E0 — endpoints/downloadCsv/envelopes を shared へ移設 (#376)
A2 entities 再建の下ごしらえ（挙動保存・移設＋import 書換のみ）:
- src/api/endpoints.ts → src/shared/api/endpoints.ts（src/api/ 空化）。
  同ディレクトリ化に伴い client は相対 import（./client）へ。
- downloadCsv を shared/lib/download.ts へ抽出（汎用 CSV DL ヘルパ）。
- ProblemDetails / ListEnvelope（transport envelope・ドメイン非依存）を
  types/index.ts から shared/api/http-types.ts へ移設。client.ts は相対 import。
- @/api/endpoints の 14 consumer を @/shared/api/endpoints へ再ポイント。
  downloadCsv を使う 4 page は @/shared/lib/download へ分離 import。

logic-diff-0 証明: endpoints.ts は rename similarity 97%（差分は header のみ）・
移設 460 行の md5 一致（本体無改変）。types/index.ts は envelope 2 型の削除のみ。
npm run check 緑（codegen:check/type-check/lint/vitest 62/knip）。

Refs #376
