+++ frontend/src/app/layout/AppLayout.tsx
@@ -2,7 +2,7 @@ import { NavLink, Link, Outlet, useNavigate } from 'react-router-dom'
 import { useQuery } from '@tanstack/react-query'
 import { useTranslation } from '@/shared/i18n/use-translation'
 import { clearToken, isAdmin } from '@/shared/api/client'
-import { getCurrentUser } from '@/api/endpoints'
+import { getCurrentUser } from '@/shared/api/endpoints'
 import { Icon } from '@/shared/ui/icon'
 import { LogoMark } from '@/shared/ui/logo-mark'
 import { KeyboardShortcuts, openShortcutsOverlay } from '@/components/keyboard'
