+++ frontend/src/pages/dunning/DunningPage.tsx
@@ -2,8 +2,9 @@ import { useState } from 'react'
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import {
   listDunningNotices, sendDunningNotice, previewDunningNotice, testSendDunningNotice, listUpstreamInvoices, type DunningStage,
-  listDunningPauses, pauseDunningNotice, resumeDunningNotice, downloadCsv, dunningNoticesExportPath,
-} from '@/api/endpoints'
+  listDunningPauses, pauseDunningNotice, resumeDunningNotice, dunningNoticesExportPath,
+} from '@/shared/api/endpoints'
+import { downloadCsv } from '@/shared/lib/download'
 import type { UpstreamInvoice } from '@/types'
 import { Badge } from '@/shared/ui/badge'
 import { Button } from '@/shared/ui/button'
